requirements.txt
streamlit==1.31.0
Pillow==10.2.0
torch==2.2.0
transformers==4.37.2
gtts==2.5.0
requests==2.31.0
accelerate==0.27.0
sentencepiece==0.2.1